Search Inventory
Adde method to database manager to search an entire table for a string.
Added method to database manager to search entre database for a string. Will most likely extend searchEntireDatabase to work for all databases in the futur(it is specific to our current design)I am too tired right now.